Changed calculator to int64 and fixed tests

+++ block.go
@@ -5,7 +5,7 @@ import "time"
 type Block struct {
 	Timestamp time.Time
 	Voters    map[VoterAddress]Voter
-	Value     float64
+	Value     int64
 }
 
 func NewBlock() *Block {
